Move hashing to more appropriate package
